+++ apina-core/src/main/kotlin/fi/evident/apina/java/model/JavaModel.kt
@@ -8,49 +8,33 @@ import java.util.*
  */
 class JavaModel {
 
-    private val classes = LinkedHashMap<String, JavaClass>()
+    private val _classes = LinkedHashMap<String, JavaClass>()
+
+    val classes: Collection<JavaClass>
+        get() = _classes.values
+
+    val annotationClasses: Sequence<JavaClass>
+        get() = _classes.values.asSequence().filter { it.isAnnotation }
 
     fun addClass(aClass: JavaClass) {
-        val old = classes.putIfAbsent(aClass.name, aClass)
+        val old = _classes.putIfAbsent(aClass.name, aClass)
         if (old != null)
             throw IllegalStateException("Class ${aClass.name} was already added")
     }
 
-    fun containsClass(name: String) = name in classes
+    fun containsClass(name: String) = name in _classes
 
-    fun findClass(name: String): JavaClass? = classes[name]
+    fun findClass(name: String): JavaClass? = _classes[name]
 
     fun findClass(type: JavaType.Basic): JavaClass? = findClass(type.name)
 
     fun findClassesWithAnnotation(annotationType: JavaType.Basic): List<JavaClass> =
-            classes.values.filter { it.hasAnnotation(annotationType) }
-
-    /**
-     * Return types for annotations implied by given annotation using same logic as Spring does.
-     * E.g. if annotation `@Foo` is itself annotated by `@Bar`, then annotating an
-     * element with `@Foo` has the same effect as annotation by `@Bar`.
-     */
-    fun findAnnotationsImpliedBy(annotationType: JavaType.Basic): Set<JavaType.Basic> {
-        val result = LinkedHashSet<JavaType.Basic>()
-
-        findAnnotationsImpliedBy(annotationType, result)
-
-        return result
-    }
-
-    private fun findAnnotationsImpliedBy(annotationType: JavaType.Basic, result: MutableSet<JavaType.Basic>) {
-        if (result.add(annotationType)) {
-
-            for (cl in classes.values)
-                if (cl.isAnnotation && cl.hasAnnotation(annotationType))
-                    findAnnotationsImpliedBy(cl.type.toBasicType(), result)
-        }
-    }
+            _classes.values.filter { it.hasAnnotation(annotationType) }
 
     inline fun <reified T : Any> isInstanceOf(type: JavaType) = isInstanceOf(type, T::class.java)
 
     fun isInstanceOf(type: JavaType, requiredType: Class<*>): Boolean {
-        val javaClass = classes[type.nonGenericClassName]
+        val javaClass = _classes[type.nonGenericClassName]
 
         if (javaClass != null) {
             return javaClass.name == requiredType.name
@@ -73,5 +57,5 @@ class JavaModel {
 
     fun isNumber(type: JavaType.Basic) = type.isPrimitiveNumber || isInstanceOf<Number>(type)
 
-    override fun toString() = classes.toString()
+    override fun toString() = _classes.toString()
 }